Move axis-tick-values lru caching into our existing `Axis`
